requirements.txt
yfinance>=0.2.0
pandas>=1.5.0
numpy>=1.24.0
lxml>=4.9.0
html5lib>=1.1
requests>=2.28.0
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
pyyaml>=6.0
akshare>=1.13.62
tushare>=1.2.89

